+++ src/main/java/kr/hhplus/be/server/infra/order/jparepository/OrderJpaRepository.java
@@ -1,7 +1,6 @@
 package kr.hhplus.be.server.infra.order.jparepository;
 
 import jakarta.persistence.LockModeType;
-import kr.hhplus.be.server.domain.coupon.entity.Coupon;
 import kr.hhplus.be.server.domain.order.entity.Order;
 import kr.hhplus.be.server.domain.product.entity.Product;
 import org.springframework.data.domain.Page;
@@ -11,10 +10,9 @@ import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.util.List;
 import java.util.Optional;
 
-public interface OrderJpaRepository extends JpaRepository<Order, Integer> {
+public interface OrderJpaRepository extends JpaRepository<Order, Long> {
 
     //주문하기
     @Lock(LockModeType.PESSIMISTIC_WRITE)
@@ -23,8 +21,7 @@ public interface OrderJpaRepository extends JpaRepository<Order, Integer> {
     //상품 한행 정보
     @Lock(LockModeType.PESSIMISTIC_WRITE)
     @Query("SELECT p FROM product p WHERE p.id = :id")
-    Optional<Product> findById(@Param("id") int id);
-
+    Optional<Product> findById(@Param("id") long id);
 
     //주문 목록 조회
     Page<Order> findByUserId(@Param("userId") int userId, Pageable pageable);
